feat(contract-deploy): add middlewares and remove playground
