+++ packages/gateway/src/middlewares/handle-error.ts
@@ -1,0 +1,19 @@
+import express from "express";
+
+export const handleErrorMiddleware = (
+  error: Error,
+  req: express.Request,
+  res: express.Response,
+  next: express.NextFunction
+) => {
+  return res.status(200).json({
+    data: null,
+    errors: [
+      {
+        message: error.message,
+        locations: [],
+        path: ''
+      },
+    ],
+  });
+};
